api/requirements.txt
aioredis==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-cache==0.1.0
fastapi-cache2==0.2.1
greenlet==3.0.3
h11==0.14.0
hiredis==2.3.2
idna==3.6
jose==1.0.0
passlib==1.7.4
pendulum==3.0.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
redis==5.0.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
time-machine==2.14.0
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0